Split up EmitLLVM.hs into smaller modules.
