+++ compiler/src/Language/Granule/Codegen/Emit/EmitableDef.hs
@@ -1,0 +1,5 @@
+module Language.Granule.Codegen.Emit.EmitableDef where
+import Language.Granule.Codegen.ClosureFreeDef
+
+type EmitableExpr = ClosureFreeExpr
+type EmitableValue = ClosureFreeValue
