+++ compiler/src/Language/Granule/Codegen/Emit/Types.hs
@@ -1,0 +1,7 @@
+module Language.Granule.Codegen.Emit.Types where
+    
+import Language.Granule.Syntax.Type as GRType
+import qualified LLVM.AST.Type as IRType
+
+type GrType = GRType.Type
+type IrType = IRType.Type
